# services/speak-gwop/speak-openapi.yaml
openapi: 3.1.0
info:
  title: speak-gwop API
  version: 0.1.0
  description: |
    Standalone agent-native TTS store.

    Flow:
    1) Create Speak identity
    2) Login for Bearer token
    3) Buy credits via Gwop invoice
    4) Claim paid credits
    5) Submit async TTS jobs
    6) Poll job and download artifact URL
servers:
  - url: http://localhost:3020
...
paths:
...
  /webhooks/gwop:
    post:
      tags: [Infra]
      summary: Receive Gwop invoice events
      description: |
        Handles `invoice.paid`, `invoice.expired`, and `invoice.canceled` transitions.
      parameters:
        - in: header
          name: x-gwop-signature
          required: false
          schema:
            type: string
          description: Required when GWOP_WEBHOOK_SECRET is configured.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GwopWebhookEvent'
      responses:
        '200':
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
    SpeakTokenHeader:
      type: apiKey
      in: header
      name: x-speak-access-token
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    GwopWebhookEvent:
      type: object
      required: [event_type, data]
      properties:
        event_type:
          type: string
          example: invoice.paid
        data:
          type: object
          required: [invoice_id]
          properties:
            invoice_id:
              type: string
            status:
              type: string